* Changed float to double in QueryResult * Rows in simpleQuery are now counted

+++ QueryResult.cc
@@ -54,9 +54,9 @@ int QueryResult::getInt(std::string colname)
 	return any_cast<int>(this->getValue(colname));
 }
 
-float QueryResult::getFloat(std::string colname)
+double QueryResult::getDouble(std::string colname)
 {
-	return any_cast<float>(this->getValue(colname));
+	return any_cast<double>(this->getValue(colname));
 }
 
 std::string QueryResult::getString(std::string colname)
